Add files via upload
I added two new columns "WeekNum" and "ActiveTime" for the ease of calculation in SQL and to avoid any issues with the syntax as I was unsure of the type of SQL compiler.
There were 2 instances where the end_time is less than the start_time in the given data. Considering those two entries as anomalies, I took the active time for that instance as 0 minutes.